data updating process v rozkladu xD

+++ iSpend/TransactionController.swift
@@ -10,7 +10,7 @@ import MapKit
 import CoreLocation
 import FirebaseFirestore
 
-class TransactionController: UIViewController, UIGestureRecognizerDelegate, CLLocationManagerDelegate, MKMapViewDelegate {
+class TransactionController: UIViewController, UIGestureRecognizerDelegate, CLLocationManagerDelegate, MKMapViewDelegate, BaseProtocol {
     @IBOutlet weak var titleLabel: UILabel!
     @IBOutlet weak var counterpartyLabelTitle: UILabel!
     @IBOutlet weak var counterpartyLabel: UILabel!
@@ -28,19 +28,20 @@ class TransactionController: UIViewController, UIGestureRecognizerDelegate, CLLo
     @IBOutlet weak var dismissButtonBackground: UIButton!
     @IBOutlet weak var optionsButton: UIButton!
     
-    let db = Firestore.firestore()
-    var listener: ListenerRegistration? = nil
+    var db = Firestore.firestore()//
+    var listener: ListenerRegistration? = nil//
     var transaction: Transaction? = nil
-    var transactions: [Transaction] = []
+    var transactions: [Transaction] = []//
+    var changedValues: [String:Any] = [:]//
     
     let locationManager = CLLocationManager()
-    var dateComponentDays = DateComponents()
-    var dateComponentMonts = DateComponents()
+    let dateFormatter = DateFormatter()//
+    var dateComponentDays = DateComponents()//
+    var dateComponentMonts = DateComponents()//
 
     var transId: Int = 0
-    var LTId: Int = Int()
-    let currency: String = "CZK"
-    let dateFormatter = DateFormatter()
+    var LTId: Int = Int()//
+    var currency: String = "CZK"//
     var hasLocation: Bool = false
     let locationButtonOnSymbolName: String = "trash.circle.fill"
     let locationButtonOffSymbolName: String = "location.circle.fill"
@@ -48,10 +49,21 @@ class TransactionController: UIViewController, UIGestureRecognizerDelegate, CLLo
     let locationNotSet: String = "no location"
     var isQuickView: Bool = false
     
-    var LMI: Double = Double()
-    var LMO: Double = Double()
-    var LWI: Double = Double()
-    var LWO: Double = Double()
+    var LMI: Double = Double()//
+    var LMO: Double = Double()//
+    var LWI: Double = Double()//
+    var LWO: Double = Double()//
+    
+    var updatedLMI: Double = Double()//
+    var updatedLMO: Double = Double()//
+    var updatedLWI: Double = Double()//
+    var updatedLWO: Double = Double()//
+    
+    var LMFromDate: Date = Date()//
+    var LMToDate: Date = Date()//
+    
+    var updatedLMFromDate: Date = Date()//
+    var updatedLMToDate: Date = Date()//
     
     override func viewWillAppear(_ animated: Bool) {
         print("STARTED LISTENNING FROM TRANSACTION DETAIL...")
@@ -60,8 +72,6 @@ class TransactionController: UIViewController, UIGestureRecognizerDelegate, CLLo
     
     override func viewDidAppear(_ animated: Bool) {
         navigationController?.interactivePopGestureRecognizer?.delegate = self
-        print("SEEING TRANSACTION ID: \(transId)")
-        print("LTId: \(LTId)")
     }
     
     override func viewWillDisappear(_ animated: Bool) {
@@ -72,18 +82,8 @@ class TransactionController: UIViewController, UIGestureRecognizerDelegate, CLLo
     override func viewDidLoad() {
         super.viewDidLoad()
         setNeedsStatusBarAppearanceUpdate()
-        backButton.isHidden = isQuickView ? true : false
-        dateFormatter.dateStyle = .medium
-        dateFormatter.timeZone = .current
-        dateFormatter.dateFormat = "d. M. yyyy"
-        priceLabel.font = UIFont.monospacedSystemFont(ofSize: 30, weight: .bold)
-        currencyLabel.text = "(\(currency))"
-        mapView.isHidden = true
-        mapView.delegate = self
-        locationManager.delegate = self
-        locationLabel.isHidden = true
-        dismissButton.isHidden = isQuickView ? false : true
-        dismissButtonBackground.isHidden = isQuickView ? false : true
+        setupView()
+        setupDateFormatter()
     }
     
     func startListening() {
@@ -101,7 +101,6 @@ class TransactionController: UIViewController, UIGestureRecognizerDelegate, CLLo
             
             let map = data["transMap"] as! Dictionary<String, Any>
             LTId = data["LTId"] as? Int ?? -1
-            print("GOT LTId: \(LTId)")
             
             if let transactionData = map[String(transId)] as? [String : Any] {
                 transaction = Transaction(counterparty: transactionData["counterparty"] as? String ?? "COUNTERPARTY ERROR", date: Date(timeIntervalSince1970: TimeInterval((transactionData["date"] as! Timestamp).seconds)), id: transactionData["id"] as? Int ?? 999999, incoming: transactionData["incoming"] as? Bool ?? false, latitude: (transactionData["location"] as! GeoPoint).latitude, longitude: (transactionData["location"] as! GeoPoint).longitude, title: transactionData["title"] as? String ?? "TITLE ERROR", total: transactionData["total"] as? Double ?? 123.45)
@@ -112,7 +111,8 @@ class TransactionController: UIViewController, UIGestureRecognizerDelegate, CLLo
             mapView.addAnnotation(location)
             mapView.setCenter(location.coordinate, animated: true)
             mapView.setRegion(MKCoordinateRegion(center: location.coordinate, latitudinalMeters: 1000, longitudinalMeters: 1000), animated: true)
-            setupView()
+            
+            updateView()
         }
     }
     
@@ -134,24 +134,44 @@ class TransactionController: UIViewController, UIGestureRecognizerDelegate, CLLo
     
     func setupView() {
         view.backgroundColor = .systemBackground
+        backButton.isHidden = isQuickView ? true : false
         backButton.tintColor = .systemBlue
         backButton.setTitleColor(.systemBlue, for: .normal)
+        
+        dismissButton.isHidden = isQuickView ? false : true
         dismissButton.tintColor = .systemGray5
+        dismissButtonBackground.isHidden = isQuickView ? false : true
         dismissButtonBackground.tintColor = .systemGray
+
         optionsButton.tintColor = .systemBlue
-        titleLabel.text = transaction?.title
+        
         titleLabel.textColor = .label
-        counterpartyLabel.text = transaction?.counterparty
         counterpartyLabel.textColor = .label
-        dateLabel.text = dateFormatter.string(from: transaction!.date)
-        priceLabel.text = String(format: "%.2f", transaction!.total)
         totalLabelTitle.textColor = .secondaryLabel
         dateLabelTitle.textColor = .secondaryLabel
         counterpartyLabelTitle.textColor = .secondaryLabel
         dateLabel.textColor = .label
+        
         priceLabel.textColor = .label
+        priceLabel.font = UIFont.monospacedSystemFont(ofSize: 30, weight: .bold)
+        
+        currencyLabel.text = "(\(currency))"
         currencyLabel.textColor = .secondaryLabel
+        
+        locationManager.delegate = self
+        locationLabel.isHidden = true
         locationLabel.textColor = .secondaryLabel
+        
+        mapView.delegate = self
+        mapView.isHidden = true
+        mapView.layer.cornerRadius = 10
+    }
+    
+    func updateView() {
+        titleLabel.text = transaction?.title
+        counterpartyLabel.text = transaction?.counterparty
+        dateLabel.text = dateFormatter.string(from: transaction!.date)
+        priceLabel.text = String(format: "%.2f", transaction!.total)
         
         if (transaction?.incoming == true) {
             symbolLabel.text = "→"
@@ -163,7 +183,14 @@ class TransactionController: UIViewController, UIGestureRecognizerDelegate, CLLo
         
         locationLabel.isHidden = (transaction?.locationEnabled())! ? false : true
         mapView.isHidden = (transaction?.locationEnabled())! ? false : true
-        mapView.layer.cornerRadius = 10
+    }
+    
+    func setupDateFormatter() {
+        dateFormatter.dateStyle = .medium
+        dateFormatter.timeZone = .current
+        dateFormatter.dateFormat = "d. M. yyyy"
+        dateComponentDays.day = -7
+        dateComponentMonts.month = -1
     }
     
     func gestureRecognizerShouldBegin(_ gestureRecognizer: UIGestureRecognizer) -> Bool
@@ -207,6 +234,7 @@ class TransactionController: UIViewController, UIGestureRecognizerDelegate, CLLo
                 db.collection("iSpend").document("UtE3HXvUEmamvjtRaDDs").getDocument { (document, error) in
                     if let document = document, document.exists {
                         let data = document.data()
+                        
                         let map = data!["transMap"]! as! Dictionary<String, Any>
                         var procItem: [String:Any] = [:]
                         transactions.removeAll()
@@ -216,37 +244,44 @@ class TransactionController: UIViewController, UIGestureRecognizerDelegate, CLLo
                             transactions.append(Transaction(counterparty: procItem["counterparty"] as? String ?? "COUNTERPARTY ERROR", date: Date(timeIntervalSince1970: TimeInterval((procItem["date"] as! Timestamp).seconds)), id: procItem["id"] as? Int ?? 999999, incoming: procItem["incoming"] as? Bool ?? false, latitude: (procItem["location"] as! GeoPoint).latitude, longitude: (procItem["location"] as! GeoPoint).longitude, title: procItem["title"] as? String ?? "TITLE ERROR", total: procItem["total"] as? Double ?? 123.45))
                         }
                         
-                        transactions.sort { (tr1, tr2) -> Bool in
-                            if (tr1.date.compare(tr2.date) == .orderedDescending) {
-                                return true
-                            } else if (tr1.date.compare(tr2.date) == .orderedAscending) {
-                                return false
-                            } else {
-                                return tr1.id > tr2.id
+                        if (!transactions.isEmpty) {
+                            print("SORTING TRANSACTIONS2")
+                            transactions.sort { (tr1, tr2) -> Bool in
+                                if (tr1.date.compare(tr2.date) == .orderedDescending) {
+                                    return true
+                                } else if (tr1.date.compare(tr2.date) == .orderedAscending) {
+                                    return false
+                                } else {
+                                    return tr1.id > tr2.id
+                                }
                             }
+                            transactions.remove(at: 0)
                         }
-                        
-                        transactions.remove(at: 0)
                         
                         if let newLastTransaction: Transaction = transactions.first {
                             let newLastTransactionId: Int = newLastTransaction.id
-                            db.collection("iSpend").document("UtE3HXvUEmamvjtRaDDs").updateData(
-                                ["LTId" : newLastTransactionId,
-                                 "transMap." + transId.description : FieldValue.delete()])
+                            if (LTId != newLastTransactionId) {
+                                print("writing to firestore7")
+                                db.collection("iSpend").document("UtE3HXvUEmamvjtRaDDs").updateData(
+                                    ["LTId" : newLastTransactionId,
+                                     "transMap." + transId.description : FieldValue.delete()])
+                                
+                                updateTotals()
+                            }
                         } else {
-                            db.collection("iSpend").document("UtE3HXvUEmamvjtRaDDs").updateData(
-                                ["LTId" : -1,
-                                 "transMap." + transId.description : FieldValue.delete()])
+                            print("NO TRANSACTIONS LEFT 1")
+                            if (LTId != -1) {
+                                print("writing to firestore8")
+                                db.collection("iSpend").document("UtE3HXvUEmamvjtRaDDs").updateData(["LTId" : -1, "LMI" : 0, "LMO" : 0, "LWI" : 0, "LWO" : 0, "nextTransactionIndex" : 0, "LMFD" : Date(), "LMTD" : Date(), "transMap." + transId.description : FieldValue.delete()])
+                            }
                         }
-                        
-                        updateTotals()
                     } else {
                         print("Document does not exist")
                     }
                 }
             } else {
                 //transaction being deleted is not the latest one
-                print("NOT DELETING THE LATEST ONE")
+                print("writing to firestore9")
                 db.collection("iSpend").document("UtE3HXvUEmamvjtRaDDs").updateData(["transMap." + transId.description : FieldValue.delete()])
             }
             
@@ -263,40 +298,6 @@ class TransactionController: UIViewController, UIGestureRecognizerDelegate, CLLo
         alert.addAction(deleteAction)
         alert.addAction(cancelAction)
         present(alert, animated: true, completion: nil)
-    }
-    
-    func getAndSortTransactions() {
-        
-    }
-    
-    func updateTotals() {
-        dateComponentDays.day = -7
-        dateComponentMonts.month = -1
-        
-        LMI = 0.0
-        LMO = 0.0
-        LWI = 0.0
-        LWO = 0.0
-        
-        for item in transactions {
-            if (item.date > Calendar.current.date(byAdding: dateComponentDays, to: Date())!) {
-                if (item.incoming == true) {
-                    LWI += item.total
-                    LMI += item.total
-                } else {
-                    LWO += item.total
-                    LMO += item.total
-                }
-            } else if (item.date > Calendar.current.date(byAdding: dateComponentMonts, to: Date())!) {
-                if (item.incoming == true) {
-                    LMI += item.total
-                } else {
-                    LMO += item.total
-                }
-            }
-        }
-        
-        db.collection("iSpend").document("UtE3HXvUEmamvjtRaDDs").updateData(["LMI" : LMI, "LMO" : LMO, "LWI" : LWI, "LWO" : LWO])
     }
     
     @IBAction func dismissButtonTapped(_ sender: UIButton) {
